backend/requirements.txt
# ── Core API framework ────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1

# ── Settings validation ───────────────────────────────────────────────────────
pydantic==2.12.5
pydantic-settings==2.13.1

# ── Vector store ──────────────────────────────────────────────────────────────
chromadb==0.5.5
sentence-transformers==3.1.1

# ── Hybrid retrieval ──────────────────────────────────────────────────────────
rank-bm25==0.2.2

# ── LLM ──────────────────────────────────────────────────────────────────────
groq==0.11.0

# ── Document processing ───────────────────────────────────────────────────────
pypdf==5.0.0
pdf2image==1.17.0
pillow==10.4.0
pytesseract==0.3.13
python-docx==1.1.2

# ── Agent orchestration (Phase 2) ─────────────────────────────────────────────
langgraph==1.0.10

# ── Async + utilities ─────────────────────────────────────────────────────────
aiofiles==23.2.1
aiosqlite==0.19.0          # async SQLite for feedback store (Phase 3)
httpx==0.27.0              # async HTTP client
tenacity==8.2.3            # retry decorator for service calls